Jest handles import statements and has 1 test for FE. Back end still passes all
