+++ __mocks__/styleMock.js
@@ -1,0 +1,1 @@
+module.exports = {}
