+++ jest.config.json
@@ -1,11 +1,20 @@
 {
   "testEnvironment": "node",
   "verbose": false,
-  "moduleFileExtensions": ["js", "jsx"],
-  "transformIgnorePatterns": [
-    "/node_modules/(?!react)"
+  "moduleFileExtensions": [
+    "js",
+    "jsx"
   ],
+  "moduleDirectories": [
+    "node_modules", "client/src"
+  ],
+  "moduleNameMapper": {
+    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
+    "\\.(css|less)$": "<rootDir>/__mocks__/styleMock.js"
+  },
   "transform": {
-    "^.+\\.js$": "babel-jest"
-  }
+    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/fileTransformer.js",
+    "^.+\\.jsx?$": "babel-jest"
+  },
+  "transformIgnorePatterns": ["/node_modules/", "\\.jsx\\.[^\\\/]+$"]
 }
