+++ test/server/dbHelper.test.js
@@ -1,7 +1,7 @@
 ////////////////////////////////////////
 const mongoose = require('mongoose');
 const { insert, getListing } = require('../../server/db/dbHelper.js');
-const tester = 'servertest'
+const tester = 'servertest';
 ////////////////////////////////////////
 // CONNECT TO MOCK DATABASE ////////////
 beforeAll(async () => {
